% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.sparsestep.R
\name{predict.sparsestep}
\alias{predict}
\alias{predict.sparsestep}
\title{Make predictions from a SparseStep model}
\usage{
\method{predict}{sparsestep}(object, newx, ...)
}
\arguments{
\item{object}{Fitted \code{sparsestep} object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to
be made.}

\item{\dots}{further argument are ignored}
}
\value{
a matrix of numerical predictions of size nobs x nlambda
}
\description{
Predicts the outcome variable for the SparseStep model for
each value of lambda supplied to the model.
}
\examples{
x <- matrix(rnorm(100*20), 100, 20)
y <- rnorm(100)
fit <- sparsestep(x, y)
yhat <- predict(fit, x)
}

