% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.sparsestep.R
\name{print.sparsestep}
\alias{print.sparsestep}
\title{Print the fitted SparseStep model}
\usage{
\method{print}{sparsestep}(obj, ...)
}
\arguments{
\item{obj}{a "sparsestep" object to print}
}
\description{
Prints a short text of a fitted SparseStep model
}
\examples{
data(diabetes)
attach(diabetes)
object <- sparsestep(x, y)
print(object)
detach(diabetes)
}

