% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparsestep-package.R
\docType{package}
\name{sparsestep-package}
\alias{sparsestep-package}
\title{SparseStep: Approximating the Counting Norm for Sparse Regularization}
\description{
In the SparseStep regression model the ordinary least-squares problem is
augmented with an approximation of the exact \eqn{\ell_0}{l[0]} pseudonorm.
This approximation is made increasingly more accurate in the SparseStep
algorithm, resulting in a sparse solution to the regression problem. See
the references for more information.
}
\section{SparseStep functions}{

The main SparseStep functions are:
\describe{
\item{\code{\link{sparsestep}}}{Fit a SparseStep model for a given range of
\eqn{\lambda} values}
\item{\code{\link{path.sparsestep}}}{Fit the SparseStep model along a path
of \eqn{\lambda} values which are generated such that a model is created at
each possible level of sparsity, or until a given recursion depth is
reached.}
}

Other available functions are:
\describe{
\item{\code{\link{plot}}}{Plot the coefficient path of the SparseStep
model.}
\item{\code{\link{predict}}}{Predict the outcome of the linear model using
SparseStep}
\item{\code{\link{coef}}}{Get the coefficients from the SparseStep model}
\item{\code{\link{print}}}{Print a short description of the SparseStep
model}
}
}
\author{
Gertjan van den Burg (author and maintainer).
}

