% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cd.sparsestep.R
\name{sparsestep.cd}
\alias{sparsestep.cd}
\title{Coordinate descent algorithm for SparseStep}
\usage{
sparsestep.cd(x, y, lambdas = NULL, epsilon = 1e-05, intercept = TRUE,
  ...)
}
\description{
Coordinate descent algorithm for SparseStep
}

