% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path.sparsestep.R
\name{sparsestep.path}
\alias{sparsestep.path}
\title{Path algorithm for the SparseStep model}
\usage{
sparsestep.path(x, y, max.depth = 10, intercept = TRUE, normalize = TRUE,
  ...)
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{response}

\item{max.depth}{maximum recursion depth}

\item{...}{further arguments to sparsestep()}
}
\description{
Fits the entire regularization path for SparseStep using a
Golden Section search.
}

