% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coef.sparsestep.R
\name{coef.sparsestep}
\alias{coef}
\alias{coef.sparsestep}
\title{Get the coefficients of a fitted SparseStep model}
\usage{
\method{coef}{sparsestep}(object, ...)
}
\arguments{
\item{object}{a \code{sparsestep} object}

\item{\dots}{further argument are ignored}
}
\value{
The coefficients of the SparseStep model (i.e. the betas). If the
model was fitted with an intercept this will be the first value in the
resulting vector.
}
\description{
Returns the coefficients of the SparseStep model.
}
\examples{
x <- matrix(rnorm(100*20), 100, 20)
y <- rnorm(100)
fit <- sparsestep(x, y)
coef(fit)
}

