% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.sparsestep.R
\name{plot.sparsestep}
\alias{plot}
\alias{plot.sparsestep}
\title{Plot the SparseStep path}
\usage{
\method{plot}{sparsestep}(x, ...)
}
\arguments{
\item{x}{a \code{sparsestep} object}

\item{\dots}{further argument to matplot}
}
\description{
Plot the coefficients of the SparseStep path
}
\examples{
x <- matrix(rnorm(100*20), 100, 20)
y <- rnorm(100)
fit <- sparsestep(x, y)
plot(fit)
pth <- path.sparsestep(x, y)
plot(pth)
}

