% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.sparsestep.R
\name{print.sparsestep}
\alias{print}
\alias{print.sparsestep}
\title{Print the fitted SparseStep model}
\usage{
\method{print}{sparsestep}(x, ...)
}
\arguments{
\item{x}{a "sparsestep" object to print}

\item{...}{further argument are ignored}
}
\description{
Prints a short text of a fitted SparseStep model
}
\examples{
x <- matrix(rnorm(100*20), 100, 20)
y <- rnorm(100)
fit <- sparsestep(x, y)
print(fit)
}

